chore: move delete head branch in their own file
