+++ mergify_engine/tests/functional/test_engine.py
@@ -83,102 +83,6 @@ class TestEngineV2Scenario(base.FunctionalTestBase):
             checks[0]["output"]["title"],
         )
 
-    def test_delete_branch(self):
-        rules = {
-            "pull_request_rules": [
-                {
-                    "name": "delete on merge",
-                    "conditions": ["base=master", "label=merge", "merged"],
-                    "actions": {"delete_head_branch": None},
-                },
-                {
-                    "name": "delete on close",
-                    "conditions": ["base=master", "label=close", "closed"],
-                    "actions": {"delete_head_branch": {}},
-                },
-            ]
-        }
-
-        self.setup_repo(yaml.dump(rules))
-
-        p1, _ = self.create_pr(base_repo="main", branch="#1-first-pr")
-        p1.merge()
-        self.wait_for("pull_request", {"action": "closed"})
-
-        p2, _ = self.create_pr(base_repo="main", branch="#2-second-pr")
-        p2.edit(state="close")
-        self.wait_for("pull_request", {"action": "closed"})
-
-        self.add_label(p1, "merge")
-        self.add_label(p2, "close")
-
-        pulls = list(self.r_o_admin.get_pulls(state="all"))
-        self.assertEqual(2, len(pulls))
-
-        branches = list(self.r_o_admin.get_branches())
-        self.assertEqual(1, len(branches))
-        self.assertEqual("master", branches[0].name)
-
-    def test_delete_branch_with_dep_no_force(self):
-        rules = {
-            "pull_request_rules": [
-                {
-                    "name": "delete on merge",
-                    "conditions": ["base=master", "label=merge", "merged"],
-                    "actions": {"delete_head_branch": None},
-                }
-            ]
-        }
-
-        self.setup_repo(yaml.dump(rules))
-
-        p1, _ = self.create_pr(base_repo="main", branch="#1-first-pr")
-        p2, _ = self.create_pr(
-            base_repo="main", branch="#2-second-pr", base="#1-first-pr"
-        )
-
-        p1.merge()
-        self.wait_for("pull_request", {"action": "closed"})
-        self.add_label(p1, "merge")
-        self.wait_for("check_run", {"check_run": {"conclusion": "success"}})
-
-        pulls = list(self.r_o_admin.get_pulls(state="all"))
-        assert 2 == len(pulls)
-
-        branches = list(self.r_o_admin.get_branches())
-        assert 3 == len(branches)
-        assert {"master", "#1-first-pr", "#2-second-pr"} == {b.name for b in branches}
-
-    def test_delete_branch_with_dep_force(self):
-        rules = {
-            "pull_request_rules": [
-                {
-                    "name": "delete on merge",
-                    "conditions": ["base=master", "label=merge", "merged"],
-                    "actions": {"delete_head_branch": {"force": True}},
-                }
-            ]
-        }
-
-        self.setup_repo(yaml.dump(rules))
-
-        p1, _ = self.create_pr(base_repo="main", branch="#1-first-pr")
-        p2, _ = self.create_pr(
-            base_repo="main", branch="#2-second-pr", base="#1-first-pr"
-        )
-
-        p1.merge()
-        self.wait_for("pull_request", {"action": "closed", "number": p1.number})
-        self.add_label(p1, "merge")
-        self.wait_for("pull_request", {"action": "closed", "number": p2.number})
-
-        pulls = list(self.r_o_admin.get_pulls(state="all"))
-        assert 2 == len(pulls)
-
-        branches = list(self.r_o_admin.get_branches())
-        assert 2 == len(branches)
-        assert {"master", "#2-second-pr"} == {b.name for b in branches}
-
     def test_assign(self):
         rules = {
             "pull_request_rules": [
